update shokri branch

+++ experiment/testing-mia/losstraj-attack.py
@@ -12,13 +12,21 @@ from utils.set_seed import set_seed
 
 
 
-dataset_dir = "datasets"
-result_dir = "results"
-configs_dir = "configs"
-dataset_name = "cifar10"
+dataset_dir = "loss_traj_test/datasets"
+result_dir = "loss_traj_test/results"
+configs_dir = "loss_traj_test/configs"
+dataset_name = "loss_traj_test/cifar10"
 batch_size = 1000
 trainset_ratio = 0.5  # meaning 50% of the training set is used for training the target model
+num_epochs = 250
 
+
+def get_index(dataset):
+    index = []
+    dataloader = DataLoader(dataset, batch_size=1, shuffle=False, num_workers=2)
+    for i, data in enumerate(dataloader):
+        index.append(i)
+    return index
 
 # define a model architecture
 class TargetModel(torch.nn.Module):
@@ -46,6 +54,7 @@ class TargetModel(torch.nn.Module):
 
 def main(testing=False):
     set_seed(0)
+    device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
     # initialize the dataset
     train_transform = T.Compose([T.RandomCrop(32, padding=4),
@@ -71,55 +80,61 @@ def main(testing=False):
                             shuffle=False, num_workers=2)
 
     # prepare the shadow set and target set and then train a target model
-
-    training_set, shadow_set = random_split(trainset, [trainset_ratio * len(trainset),
-                                                       len(trainset) - trainset_ratio * len(trainset)])
+    train_len = int(len(trainset) * trainset_ratio)
+    shadow_len = len(trainset) - train_len
+    training_set, shadow_set = random_split(trainset, [train_len, shadow_len])
 
     # defining a target model
     """NOTE: again, we are allowing user to define their own target model"""
     target_model = TargetModel()
 
-    # training the target model
-    total_loss = 0
-    correct_predictions = 0
-    total_samples = 0
-    training_loader = DataLoader(training_set, batch_size=batch_size, shuffle=True, num_workers=2)
-    target_model.train()
-    optimizer = torch.optim.Adam(target_model.parameters(), lr=0.001)
-    criterion = torch.nn.CrossEntropyLoss()
-    for epoch in tqdm(range(100)):
-        for i, data in enumerate(training_loader):
-            inputs, labels = data
-            optimizer.zero_grad()
-            outputs = target_model(inputs)
-            loss = criterion(outputs, labels)
-            loss.backward()
-            optimizer.step()
+    if os.path.exists(os.path.join(result_dir, "target_model.pth")):
+        target_model.load_state_dict(torch.load(os.path.join(result_dir, "target_model.pth")))
+    else:
+        # training the target model
+        target_model.to(device)
+        correct_predictions = 0
+        total_samples = 0
+        training_loader = DataLoader(training_set, batch_size=batch_size, shuffle=True, num_workers=2)
+        target_model.train()
+        optimizer = torch.optim.Adam(target_model.parameters(), lr=0.005, weight_decay=0.0005)
+        criterion = torch.nn.CrossEntropyLoss()
 
-            # Calculate the accuracy for this batch
-            _, predicted = torch.max(outputs, 1)
-            correct_predictions += (predicted == labels).sum().item()
-            total_samples += labels.size(0)
+        for epoch in range(num_epochs):
+            for i, data in enumerate(training_loader):
+                inputs, labels = data
+                inputs, labels = inputs.to(device), labels.to(device)
+                optimizer.zero_grad()
+                outputs = target_model(inputs)
+                loss = criterion(outputs, labels)
+                loss.backward()
+                optimizer.step()
 
-            # Print batch-level loss and accuracy
-            tqdm.write(
-                f"Epoch {epoch + 1}, Batch {i + 1}, Loss: {loss.item():.4f}, Accuracy: {(correct_predictions / total_samples) * 100:.2f}%")
+                # Calculate the accuracy for this batch
+                _, predicted = torch.max(outputs, 1)
+                correct_predictions += (predicted == labels).sum().item()
+                total_samples += labels.size(0)
 
+            accuracy = (correct_predictions / total_samples) * 100
+            if epoch % 20 == 0:
+                print(f"Epoch {epoch + 1},\tAccuracy: {accuracy:.2f}%,\tLoss: {loss.item():.4f}")
 
+        # Calculate and print the accuracy at the end of the epoch
+        accuracy = (correct_predictions / total_samples) * 100
+        print(f"End of Epoch {epoch + 1}, Accuracy: {accuracy:.2f}%")
 
-    # Calculate and print the accuracy at the end of the epoch
-    accuracy = (correct_predictions / total_samples) * 100
-    print(f"End of Epoch {epoch + 1}, Accuracy: {accuracy:.2f}%")
-
-    # save the target model
-    torch.save(target_model.state_dict(), os.path.join(result_dir, "target_model.pth"))
+        # save the target model
+        if not os.path.exists(result_dir):
+            os.makedirs(result_dir)
+        torch.save(target_model.state_dict(), os.path.join(result_dir, "target_model.pth"))
 
     # initialize the target model access
     target_model_access = losstraj_mia.LosstrajModelAccess(target_model, mia_base.ModelAccessType.WHITE_BOX)
 
     # initialize the attack
-    merlin_attack_config = losstraj_mia.LosstrajAuxiliaryInfo(None)  # None would result in default config
-    attack = losstraj_mia.LosstrajAttack(target_model_access, merlin_attack_config)
+    config = {'save_path': "loss_traj_test/mia_files"}
+    losstraj_attack_config = losstraj_mia.LosstrajAuxiliaryInfo(config)  # None would result in default config
+    attack = losstraj_mia.LosstrajAttack(target_model_access, losstraj_attack_config)
 
     attack.prepare(shadow_set)
 
@@ -127,11 +142,22 @@ def main(testing=False):
     attack_test = ConcatDataset([training_set, testset])
     y_membership = np.concatenate((np.ones(len(training_set)), np.zeros(len(training_set))))
 
+    index = get_index(attack_test)
+
     pred_membership = attack.infer(attack_test)
 
     # calculate the accuracy
     accuracy = np.sum(y_membership == pred_membership) / len(y_membership)
     print(f"Accuracy: {accuracy:.2f}%")
+
+    # save the results
+    if not os.path.exists(result_dir + "/" + "attack result"):
+        os.makedirs(result_dir + "/" + "attack result")
+    np.save(os.path.join(result_dir + "/" + "attack result", "y_membership.npy"), y_membership)
+    np.save(os.path.join(result_dir + "/" + "attack result", "pred_membership.npy"), pred_membership.cpu().detach().numpy())
+    np.save(os.path.join(result_dir + "/" + "attack result", "index.npy"), index)
+
+
 
 
 if __name__ == "__main__":
